pyproject.toml
[tool.poetry]
name = "non-parametric-ml"
version = "0.1.0"
description = ""
authors = ["Matej <matouhrin@gmail.com>"]
readme = "README.md"
packages = [{include = "non_parametric_ml"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
pandas = "^2.2.0"
polars = "^0.20.9"
ipykernel = "^6.29.2"
seaborn = "^0.13.2"
plotly = "^5.19.0"
nbformat = "^5.9.2"
lightning = "^2.2.0.post0"
torchvision = "^0.17.0"
tensorboard = "^2.16.2"
yfinance = "^0.2.36"
pyarrow = "^15.0.0"
empiricaldist = "^0.7.2"
scipy = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
